add waves and fix bugs

+++ contracts/ethereum/contracts/ISubscription.sol
@@ -1,5 +1,5 @@
 pragma solidity >=0.4.21 <0.7.0;
 
 interface ISubscription {
-    function attachData(uint256 data) external;
+    function attachData(uint64 data) external;
 }
